# fix(next/image): reduce avif sharp effort from 4 to 3 for faster encoding