--- a/packages/next/src/server/image-optimizer.ts
+++ b/packages/next/src/server/image-optimizer.ts
@@ -549,6 +549,7 @@ export async function optimizeImage({
     const avifQuality = quality - 20
     transformer.avif({
       quality: Math.max(avifQuality, 1),
+      chromaSubsampling: '4:2:0', // same as webp
     })
   } else if (contentType === WEBP) {
     transformer.webp({ quality })