@styfle 2024-11-21
@lovell Thanks!

I have been experimenting with the more of the options and I think I'm happy using `effort: 3` (along with the existing `quality - 20`.

This seems to be consistently fast, even faster than chroma subsampling at effort 4, and the output is a similar size.


Using `w=1850` and `q=75`:
```
Before: 3_148ms and 153_364 bytes
 After: 1_408ms and 158_862 bytes
```

Using `w=1500` and `q=75`:
```
Before: 640ms and 95_810 bytes
 After: 175ms and 99_127 bytes
```

Using `w=1200` and `q=75`:
```
Before: 502ms and 74_738 bytes
 After: 146ms and 77_955 bytes
```

While these are a little larger, they're still smaller than webp at quality 75.